backend/django/requirements.txt
Django==3.0.4
uwsgi==2.0.18
mysqlclient==1.4.2.post1
djangorestframework==3.11.0
django-filter==2.0.0
djoser==2.0.3
django-cors-headers==3.2.1
django-environ==0.4.5

djangorestframework-simplejwt==4.4.0

beautifulsoup4==4.9.0
Markdown==3.2.1

# file
Pillow==7.1.1
django-cleanup==4.0.0
django-imagekit==4.0.2

# dev
factory-boy==2.12.0
mock==4.0.1
pytest==5.3.5
pytest-django==3.8.0
pytest-cov==2.8.1

